refactor: move all dependency tags to `tag.go`
Clean up Bob codebase conventions.
