+++ core/generated.go
@@ -28,16 +28,6 @@ import (
 	"github.com/ARM-software/bob-build/internal/warnings"
 )
 
-var (
-	GeneratedHeadersTag       = DependencyTag{name: "generated_headers"}
-	ExportGeneratedHeadersTag = DependencyTag{name: "export_generated_headers"}
-	GeneratedSourcesTag       = DependencyTag{name: "generated_sources"}
-	GeneratedTag              = DependencyTag{name: "generated_dep"}
-	HostToolBinaryTag         = DependencyTag{name: "host_tool_bin"}
-	FilegroupTag              = DependencyTag{name: "filegroup"}
-	ImplicitSourcesTag        = DependencyTag{name: "implicit_srcs"}
-)
-
 // Return a list of headers generated by this module with full paths
 func getHeadersGenerated(m dependentInterface) []string {
 	return append(m.outputs(), m.implicitOutputs()...)
